src/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1; python_version >= '3.6'
asyncpg==0.23.0
click==8.0.1; python_version >= '3.6'
databases==0.4.3
fastapi==0.68.0
h11==0.12.0; python_version >= '3.6'
psycopg2==2.9.1
pydantic==1.8.2
python-decouple==3.4
sqlalchemy==1.3.24
starlette==0.14.2; python_version >= '3.6'
typing-extensions==3.10.0.0
uvicorn==0.14.0
